+++ Photon/src/SPI-WITH-MSP430/SPI-WITH-MSP430.h
@@ -59,7 +59,7 @@ extern float Color;
 extern float Desired_Speed;
 
 extern char robotPlay; //  A “1” indicates the robot has to start running. A “0” indicates a stop signal.
-extern int timeStamp; // Timestamp of the latest SPI data
+extern unsigned long int timeStamp; // Timestamp of the latest SPI data
 
 extern char sharpestCurve; // Tightest curve radius in cm
 extern float cruiseKp; // Proportional gain of the lineCruiser
